requirements.txt
abydos==0.5.0
azure-cognitiveservices-vision-computervision==0.9.0
language-tool-python==2.7.1
Pillow==9.2.0
textblob==0.17.1
streamlit==1.12.0
streamlit-lottie==0.0.3
SpeechRecognition==3.8.1
pyttsx3==2.90
eng-to-ipa==0.0.2
msrest==0.6.21
requests==2.26.0
pandas==1.3.3
scikit-learn==1.0.2
numpy==1.21.0
